+++ src/main/java/com/jcloisterzone/game/phase/BazaarPhase.java
@@ -4,7 +4,11 @@ import java.util.ArrayList;
 
 import com.jcloisterzone.Player;
 import com.jcloisterzone.PointCategory;
+import com.jcloisterzone.action.ActionsState;
+import com.jcloisterzone.action.BazaarSelectTileAction;
 import com.jcloisterzone.board.Tile;
+import com.jcloisterzone.board.TileDefinition;
+import com.jcloisterzone.board.TilePackState;
 import com.jcloisterzone.event.BazaarAuctionEndEvent;
 import com.jcloisterzone.event.BazaarMakeBidEvent;
 import com.jcloisterzone.event.BazaarSelectBuyOrSellEvent;
@@ -14,11 +18,17 @@ import com.jcloisterzone.game.CustomRule;
 import com.jcloisterzone.game.Game;
 import com.jcloisterzone.game.Snapshot;
 import com.jcloisterzone.game.capability.BazaarCapability;
+import com.jcloisterzone.game.capability.BazaarCapabilityModel;
 import com.jcloisterzone.game.capability.BazaarItem;
 import com.jcloisterzone.game.state.CapabilitiesState;
+import com.jcloisterzone.game.state.GameState;
+import com.jcloisterzone.game.state.GameState.Flag;
 import com.jcloisterzone.ui.GameController;
 import com.jcloisterzone.wsio.WsSubscribe;
 import com.jcloisterzone.wsio.message.PassMessage;
+
+import io.vavr.Tuple2;
+import io.vavr.collection.Queue;
 
 public class BazaarPhase extends ServerAwarePhase {
 
@@ -31,67 +41,64 @@ public class BazaarPhase extends ServerAwarePhase {
         return capabilities.contains(BazaarCapability.class);
     }
 
-    @Override
-    public Player getActivePlayer() {
-        Player bidding =  bazaarCap.getBazaarBiddingPlayer();
-        return bidding == null ? bazaarCap.getBazaarTileSelectingPlayer() : bidding;
-    }
+//    @Override
+//    public Player getActivePlayer() {
+//        Player bidding =  bazaarCap.getBazaarBiddingPlayer();
+//        return bidding == null ? bazaarCap.getBazaarTileSelectingPlayer() : bidding;
+//    }
 
     @Override
-    public void enter() {
-        if (!isBazaarTriggered()) {
-            next();
+    public void enter(GameState state) {
+        if (!state.hasFlag(Flag.BAZAAR_AUCTION)) {
+            next(state);
             return;
         }
-        Player p = game.getNextPlayer();
-        bazaarCap.setBazaarTileSelectingPlayer(p);
-        int size = game.getAllPlayers().length();
-        ArrayList<BazaarItem> supply = new ArrayList<BazaarItem>(size);
+
+        int size = state.getPlayers().length();
+        TilePackState tilePack = state.getTilePack();
+
+        if (tilePack.size() < size) {
+            next(state);
+            return;
+        }
+
+        Queue<BazaarItem> supply = Queue.empty();
+
         for (int i = 0; i < size; i++) {
-            Tile t = getTilePack().drawTile(game.getRandom().nextInt(getTilePack().size()));
-            supply.add(new BazaarItem(t));
+            int rndIndex = game.getRandom().nextInt(tilePack.size());
+            Tuple2<TileDefinition, TilePackState> t = tilePack.drawTile(rndIndex);
+            state = state.setTilePack(t._2);
+            supply = supply.append(new BazaarItem(t._1, 0, null, null));
         }
-        bazaarCap.setBazaarSupply(supply);
-        toggleClock(getActivePlayer());
-        game.post(new BazaarSelectTileEvent(getActivePlayer(), supply));
+
+        Player player = state.getTurnPlayer().getNextPlayer(state);
+        BazaarCapabilityModel model = new BazaarCapabilityModel(supply, null, player);
+
+        state = state.setCapabilityModel(BazaarCapability.class, model);
+
+        toggleClock(player);
+        BazaarSelectTileAction action = new BazaarSelectTileAction(supply.toLinkedSet());
+        state = state.setPlayerActions(
+            new ActionsState(player, action, false)
+        );
+        promote(state);
     }
 
+//
+//
+//    private boolean isBazaarTriggered() {
+//        if (!bazaarCap.isBazaarTriggered()) return false;
+//        if (getTilePack().size() < game.getAllPlayers().length()) return false; //there isn't one tile for each player available
+//        if (bazaarCap.getBazaarSupply() != null) return false;
+//        return true;
+//    }
 
-    @Override
-    public void loadGame(Snapshot snapshot) {
-        setEntered(true); //avoid call enter on load phase to this phase switch
-        Player selecting = bazaarCap.getBazaarTileSelectingPlayer();
-        if (selecting != null) {
-            Player bidding = bazaarCap.getBazaarBiddingPlayer();
-            BazaarItem currentItem = bazaarCap.getCurrentBazaarAuction();
-            int supplyIdx = bazaarCap.getBazaarSupply().indexOf(currentItem);
-
-            if (bidding == null) {
-                toggleClock(getActivePlayer());
-                game.post(new BazaarSelectTileEvent(getActivePlayer(), bazaarCap.getBazaarSupply()));
-            } else if (selecting == bidding) {
-                toggleClock(bidding);
-                game.post(new BazaarSelectBuyOrSellEvent(bidding, currentItem, supplyIdx));
-            } else {
-                toggleClock(bidding);
-                game.post(new BazaarMakeBidEvent(bidding, currentItem, supplyIdx));
-            }
-        }
-    }
-
-    private boolean isBazaarTriggered() {
-        if (!bazaarCap.isBazaarTriggered()) return false;
-        if (getTilePack().size() < game.getAllPlayers().length()) return false; //there isn't one tile for each player available
-        if (bazaarCap.getBazaarSupply() != null) return false;
-        return true;
-    }
-
-    private boolean canPlayerBid(Player p) {
-        for (BazaarItem bi : bazaarCap.getBazaarSupply()) {
-            if (bi.getOwner() == p) return false;
-        }
-        return true;
-    }
+//    private boolean canPlayerBid(Player p) {
+//        for (BazaarItem bi : bazaarCap.getBazaarSupply()) {
+//            if (bi.getOwner() == p) return false;
+//        }
+//        return true;
+//    }
 
     @Override
     public void bazaarBid(Integer supplyIndex, Integer price) {